Pre-release v27

+++ test.sh
@@ -155,7 +155,7 @@ then
 fi
 
 echo "[INFO] Refreshing static files"
-rm -rf "$STATIC_DIR"*     # keeps top directory
+[ -d "$STATIC_DIR" ] && rm -rf "$STATIC_DIR"*     # keeps top directory
 COLLECT=$(./manage.py collectstatic --link)
 RES=$?
 if [ $RES -ne 0 ]
@@ -166,7 +166,7 @@ fi
 
 # create a link from /tmp/static to the actual static dir
 # used to load static content from html written by e2e_open_in_browser()
-rm -rf "$TMP_HTML"
+[ -d "$TMP_HTML" ] && rm -rf "$TMP_HTML"
 mkdir -p "$TMP_HTML"
 ln -s "$STATIC_DIR" "$TMP_HTML/static"
 
@@ -196,6 +196,7 @@ then
 
     echo "[INFO] Running migrations and performing run with nodebug"
     # ..and add coverage with no-debug
+    # -v 2 shows progress of migrations
     python3 "${PY_OPTS[@]}" -u "${PYCOV[@]}" ./manage.py test --keepdb --noinput --settings=$SETTINGS_AUTOTEST_NODEBUG -v 2 Plein.tests.tests.TestPlein.test_quick &>>"$LOG"
     RES=$?
     [ $RES -eq 0 ] || ABORTED=1
@@ -203,8 +204,8 @@ then
 
     echo "[INFO] Running manage.py exit test"
     # trigger diff that generates exit code
-    ./manage.py shell -c 'from ImportCRM.models import ImportLimieten as L; l = L.objects.first(); l.max_club_changes=1; l.save()'
-    python3 "${PY_OPTS[@]}" -u "${PYCOV[@]}" ./manage.py diff_crm_jsons ./ImportCRM/test-files/testfile_19.json ./ImportCRM/test-files/testfile_23.json >/dev/null
+    ./manage.py shell -c 'from ImportCRM.models import ImportLimieten as L; l = L.objects.first(); l.max_club_changes=1; l.save()' &>>"$LOG"
+    python3 "${PY_OPTS[@]}" -u "${PYCOV[@]}" ./manage.py diff_crm_jsons ./ImportCRM/test-files/testfile_19.json ./ImportCRM/test-files/testfile_23.json &>/dev/null
     RES=$?
     [ $RES -ne 0 ] || ABORTED=1
     # echo "[DEBUG] Debug run result: $RES --> ABORTED=$ABORTED"
@@ -259,8 +260,7 @@ powerprofilesctl set performance
 if [ $ABORTED -eq 0 ]
 then
     echo "[INFO] Starting main test run" >>"$LOG"
-    # echo "[DEBUG] python3 ${PY_OPTS[*]} -u ${PYCOV[*]} ./manage.py test --keepdb --settings=$SETTINGS_AUTOTEST -v 2 ${FOCUS_ARGS[*]}"
-    python3 "${PY_OPTS[@]}" -u "${PYCOV[@]}" ./manage.py test --keepdb --settings=$SETTINGS_AUTOTEST -v 2 "${FOCUS_ARGS[@]}" &>>"$LOG"
+    python3 "${PY_OPTS[@]}" -u "${PYCOV[@]}" ./manage.py test --keepdb --settings=$SETTINGS_AUTOTEST --exclude-tag=browser -v 2 "${FOCUS_ARGS[@]}" &>>"$LOG"
     RES=$?
     #echo "[DEBUG] Run result: $RES --> ABORTED=$ABORTED"
     [ $RES -eq 3 ] && ABORTED=1
@@ -278,6 +278,18 @@ then
         read -ra HTML_FILES < <(ls -1tr "$TMP_HTML"/*html)   # sorted by creation time
         firefox "${HTML_FILES[@]}" &
     fi
+fi
+
+if [ $ABORTED -eq 0 ]
+then
+    echo "[INFO] Starting browser test run" >>"$LOG"
+    python3 "${PY_OPTS[@]}" -u "${PYCOV[@]}" ./manage.py test --keepdb --settings=$SETTINGS_AUTOTEST --tag=browser -v 2 "${FOCUS_ARGS[@]}" &>>"$LOG"
+    RES=$?
+    #echo "[DEBUG] Run result: $RES --> ABORTED=$ABORTED"
+    [ $RES -eq 3 ] && ABORTED=1
+
+    echo >>"$LOG"
+    echo "[INFO] Finished browser test run" >>"$LOG"
 fi
 
 # stop showing the additions to the logfile, because the rest is less interesting
@@ -332,7 +344,7 @@ kill $PID_WEBSIM3
 wait $PID_WEBSIM3 2>/dev/null
 
 # cleanup test data directories
-rm -rf "$TEST_DIR"
+[ -d "$TEST_DIR" ] && rm -rf "$TEST_DIR"
 
 ASK_LAUNCH=0
 COVERAGE_RED=0
@@ -343,7 +355,7 @@ then
     echo "[INFO] Generating reports" | tee -a "$LOG"
 
     # delete old coverage report
-    rm -rf "$REPORT_DIR" &>>"$LOG"
+    [ -d "$REPORT_DIR" ] && rm -rf "$REPORT_DIR" &>>"$LOG"
 
     if [ -z "$FOCUS" ] || [ $FORCE_FULL_COV -ne 0 ]
     then
